+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Compute shader variant view reports kernel thread count for Unity 2021.2 and above
 * View-specific descriptions
 * AudioClip diagnostics
+* Material list view in shader module
 
 ### Changed
 * Improved issue description and suggestion strings, to make them clearer and more consistent